tests: merge ainic tests into test_ainic.py
Combine the compile-time support check (formerly test_ainic_compiled.py)
and the RDMA performance-track check (formerly test_ainic_perf.py) into a
single TestAINIC class in test_ainic.py. Markers are moved to method level
so the build-artifact check stays hardware-independent (rocprof_binary +
amdsmi_min_version) while the performance test keeps its ainic/network/rocpd
requirements. CTest names are preserved: ainic-settings-present and
ainic-performance-tracks.
